+++ DeZogPlugin/Changelog.md
@@ -1,9 +1,9 @@
 # Changelog
 
 ## 2.4.0
-- Support for DZRP 2.2.0. Added/changed DZRP commands:
+- Support for DZRP 2.2.0. Added DZRP commands:
   - CMD_GET_SUPPORTED_COMMANDS
-  - CMD_WRITE_MEM/CMD_READ_MEM
+  - CMD_WRITE_BANK_MEM/CMD_READ_BANK_MEM
 
 ## 2.3.0
 - Added DZRP commands: